+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/BaseMixedRealityProfileInspector.cs
@@ -1,12 +1,10 @@
 ﻿// Copyright (c) XRTK. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.﻿
 
-using System;
 using UnityEditor;
 using UnityEngine;
 using XRTK.Definitions;
 using XRTK.Inspectors.Extensions;
-using XRTK.Inspectors.PropertyDrawers;
 using XRTK.Inspectors.Utilities;
 using XRTK.Services;
 using XRTK.Utilities.Async;
@@ -47,58 +45,8 @@ namespace XRTK.Inspectors.Profiles
             EditorGUILayout.Space();
         }
 
-        /// <summary>
-        /// Creates a new <see cref="BaseMixedRealityProfile"/> instance and sets it to the <see cref="SerializedProperty"/>.
-        /// </summary>
-        /// <param name="parentProfile"></param>
-        /// <param name="property"></param>
-        /// <param name="profileType"></param>
-        protected static BaseMixedRealityProfile CreateNewProfileInstance(BaseMixedRealityProfile parentProfile, SerializedProperty property, Type profileType = null)
-        {
-            ScriptableObject instance;
-
-            if (profileType == null)
-            {
-                if (!string.IsNullOrWhiteSpace(property.type))
-                {
-                    var profileTypeName = property.type?.Replace("PPtr<$", string.Empty).Replace(">", string.Empty);
-                    instance = CreateInstance(profileTypeName);
-                }
-                else
-                {
-                    Debug.LogError("No property type found!");
-                    return null;
-                }
-            }
-            else
-            {
-                instance = CreateInstance(profileType);
-            }
-
-            Debug.Assert(Selection.activeObject != null);
-            var assetPath = AssetDatabase.GetAssetPath(Selection.activeObject);
-            var newProfile = instance.CreateAsset(assetPath) as BaseMixedRealityProfile;
-            Debug.Assert(newProfile != null);
-            newProfile.ParentProfile = parentProfile;
-            property.objectReferenceValue = newProfile;
-            return newProfile;
-        }
-
-        /// <summary>
-        /// Clones an instance of a <see cref="BaseMixedRealityProfile"/> and sets it to the <see cref="SerializedProperty"/>.
-        /// </summary>
-        /// <param name="parentProfile"></param>
-        /// <param name="property"></param>
-        /// <param name="renderedProfile"></param>
-        protected static void CloneProfileInstance(BaseMixedRealityProfile parentProfile, SerializedProperty property, BaseMixedRealityProfile renderedProfile)
-        {
-            profileSource = renderedProfile;
-            var newProfile = CreateNewProfileInstance(parentProfile, property, renderedProfile.GetType());
-            PasteProfileValuesDelay(newProfile);
-        }
-
         [MenuItem("CONTEXT/BaseMixedRealityProfile/Create Clone from Profile Values", false, 0)]
-        protected static async void CreateCloneProfile()
+        private static async void CreateCloneProfile()
         {
             profileSource = currentlySelectedProfile;
             var newProfile = CreateInstance(currentlySelectedProfile.GetType().ToString());
